requirements.txt
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
confluent-kafka==2.10.0
contourpy==1.3.1
cryptography==44.0.2
cx_Oracle==8.3.0
cycler==0.12.1
Flask==3.1.1
fonttools==4.55.3
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
httplib2==0.30.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.3
numpy==2.2.0
oauthlib==3.3.1
oracledb==3.1.0
packaging==24.2
pillow==11.0.0
proto-plus==1.26.1
protobuf==6.32.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
